+++ examples/enterprise/index.js
@@ -1,16 +1,32 @@
 /* eslint-disable no-console, import/no-unresolved */
 import fs from 'node:fs';
+import minimist from 'minimist';
 import yaml from 'yaml';
 import { App } from 'octokit';
 import Knuff from '@acuminous/knuff';
 import GitHubDriver from '@acuminous/knuff-github-driver';
 
-const GITHUB_APP_ID = process.env.GITHUB_APP_ID;
-const GITHUB_INSTALLATION_ID = process.env.GITHUB_INSTALLATION_ID;
-const PRIVATE_KEY_PATH = process.env.PRIVATE_KEY_PATH;
-const PATH_TO_REMINDERS = process.env.PATH_TO_REMINDERS || 'reminders.yaml';
+const argv = minimist(process.argv.slice(2), {
+  booleans: ['pretend'],
+  default: {
+    appId: process.env.GITHUB_APP_ID,
+    privateKey: 'private-key.pem',
+    installationId: process.env.GITHUB_INSTALLATION_ID,
+    reminders: 'reminders.yaml',
+    pretend: false,
+    now: Date.now(),
+  },
+});
+
+const appId = argv.appId;
+const privateKey = fs.readFileSync(argv.privateKey, 'utf8');
+const installationId = argv.installationId;
+const pathToReminders = argv.reminders;
+const pretend = argv.pretend;
+const now = () => new Date(argv.now).getTime();
 
 const config = {
+  pretend,
   repositories: {
     'acuminous/knuff': {
       owner: 'acuminous',
@@ -22,14 +38,13 @@ const config = {
 
 (async () => {
   try {
-    const privateKey = fs.readFileSync(PRIVATE_KEY_PATH, 'utf8');
-    const app = new App({ appId: GITHUB_APP_ID, privateKey });
-    const octokit = await app.getInstallationOctokit(GITHUB_INSTALLATION_ID);
+    const app = new App({ appId, privateKey });
+    const octokit = await app.getInstallationOctokit(installationId);
     const drivers = { github: new GitHubDriver(octokit.rest) };
-    const knuff = new Knuff(config, drivers)
+    const knuff = new Knuff(config, drivers, now)
       .on('error', console.error)
       .on('progress', console.log);
-    const reminders = yaml.parse(fs.readFileSync(PATH_TO_REMINDERS, 'utf8'));
+    const reminders = yaml.parse(fs.readFileSync(pathToReminders, 'utf8'));
     const stats = await knuff.process(reminders);
     console.log(`Successfully processed ${stats.reminders} reminders`);
   } catch (error) {